python_app/requirements.txt
Flask>=3.0.0
Jinja2>=3.1.0
Werkzeug>=3.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=42.0.0
bcrypt>=4.1.0
pymongo>=4.6.0
celery>=5.4.0
redis>=5.0.0
gunicorn>=22.0.0